test: Clean up
