+++ pipeline/Jenkinsfile
@@ -67,15 +67,15 @@ parameters
 
             try {
                 if(env.TESTS_TO_RUN){
-                    sh "py.test -k '${env.TESTS_TO_RUN}' --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results"
+                    sh "py.test -k '${env.TESTS_TO_RUN}' -n '${env.NO_OF_PARALLEL_RUNNERS}' --dist=loadfile --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results"
                 } else {
-                sh '''
+                    sh """
                     py.test -k 'smoke' --reruns 1 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
-                    if [ $? -eq 0 ]
+                    if [ \$? -eq 0 ]
                     then
-                         py.test -k 'regression' -n 3 --dist=loadfile --reruns 2 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
+                         py.test -k 'regression' -n '${env.NO_OF_PARALLEL_RUNNERS}' --dist=loadfile --reruns 2 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
                     fi
-                '''
+                    """
                 }
             } catch (error) {
                     archiveArtifacts artifacts: "ui_automation_tests/allure-results/**.*"
