+++ ui_automation_tests/features/hmrc_query.feature
@@ -8,6 +8,6 @@ Feature: Add a HMRC query
     When I go to the case list page
     Then I should see my case in the cases list
     And I should see my case SLA
-    When I go to the case
+    When I go to application previously created
     And I go to the documents tab
     Then I see my autogenerated application form
