+++ ui_automation_tests/features/view_mod_clearance.feature
@@ -8,7 +8,7 @@ Feature: View an MOD Clearance
     When I go to the case list page
     Then I should see my case in the cases list
     And I should see my case SLA
-    When I go to the case
+    When I go to application previously created
     Then I see the case page
     When I go to the documents tab
     Then I see my autogenerated application form
@@ -20,7 +20,7 @@ Feature: View an MOD Clearance
     When I go to the case list page
     Then I should see my case in the cases list
     And I should see my case SLA
-    When I go to the case
+    When I go to application previously created
     Then I see the case page
     When I go to the documents tab
     Then I see my autogenerated application form
@@ -32,7 +32,7 @@ Feature: View an MOD Clearance
     When I go to the case list page
     Then I should see my case in the cases list
     And I should see my case SLA
-    When I go to the case
+    When I go to application previously created
     Then I see the case page
     When I go to the documents tab
     Then I see my autogenerated application form
