+++ ui_automation_tests/features/view_transhipment_application.feature
@@ -11,7 +11,7 @@ Feature: I want to view the case details of a case
     When I go to the case list page
     Then I should see my case in the cases list
     And I should see my case SLA
-    When I go to the case
+    When I go to application previously created
     Then I see the case page
     When I go to the documents tab
     Then I see my autogenerated application form
