+++ ui_automation_tests/step_defs/test_filter_cases.py
@@ -78,3 +78,9 @@ def i_filter_by_goods_related_description(driver, context):
 @when("I apply filters")
 def i_apply_filters(driver, context):
     CaseListPage(driver).click_apply_filters_button()
+
+
+@when(parsers.parse('filter status has been changed to "{status}"'))  # noqa
+def filter_status_change(driver, context, status):  # noqa
+    CaseListPage(driver).select_filter_status_from_dropdown(status)
+    CaseListPage(driver).click_apply_filters_button()
